backend/requirements.txt
absl-py==1.4.0
accelerate==0.19.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.2
bitsandbytes==0.39.0
bleach==6.0.0
blinker==1.6.2
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clean-fid==0.1.35
click==8.1.3
cmake==3.26.3
colab-convert==2.0.5
comm==0.1.3
contourpy==1.0.7
cuda-python==12.1.0
cycler==0.11.0
Cython==0.29.35
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
devtools==0.11.0
diffusers @ git+https://github.com/ShivamShrirao/diffusers@2c298300756fa7d8bf644852cffaebc5072f11f6
einops==0.6.1
exceptiongroup==1.1.1
executing==1.2.0
facexlib==0.3.0
fastapi==0.95.2
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.12.0
filterpy==1.4.5
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JSON==0.4.0
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.3
-e git+https://github.com/TencentARC/GFPGAN.git@2eac2033893ca7f427f4035d80fe95b92649ac56#egg=gfpgan
google-auth==2.19.1
google-auth-oauthlib==1.0.0
gradio==3.33.1
gradio-client==0.2.5
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
imageio==2.30.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonmerge==1.9.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.0
jupyter-server==2.6.0
jupyter-server-terminals==0.4.4
jupyterlab==4.0.1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.22.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kornia==0.6.12
linkify-it-py==2.0.2
lit==16.0.5.post0
llvmlite==0.40.1rc1
lmdb==1.4.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.3.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook-shim==0.2.3
numba==0.57.0
numexpr==2.8.4
numpngw==0.1.2
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.9.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==1.7.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
realesrgan==0.3.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tb-nightly==2.14.0a20230613
tensorboard==2.13.0
tensorboard-data-server==0.7.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trampoline==0.1.2
transformers==4.28.1
triton==2.0.0
typing-extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.2.0
urllib3==1.26.16
uvicorn==0.22.0
validators==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.4
widgetsnbextension==4.0.7
xformers==0.0.20
yapf==0.40.0
yarl==1.9.2
zipp==3.15.0
